backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.32
pymysql==1.1.1
pyodbc==4.0.39
python-dotenv==1.0.0
